bring generateSessionId function to sessions.ts
