--- ui/desktop/src/utils/session.ts
@@ -1,14 +1,0 @@
-/**
- * Generate a session ID in the format yyyymmdd_hhmmss
- */
-export function generateSessionId(): string {
-  const now = new Date();
-  const year = now.getFullYear();
-  const month = String(now.getMonth() + 1).padStart(2, '0');
-  const day = String(now.getDate()).padStart(2, '0');
-  const hours = String(now.getHours()).padStart(2, '0');
-  const minutes = String(now.getMinutes()).padStart(2, '0');
-  const seconds = String(now.getSeconds()).padStart(2, '0');
-
-  return `${year}${month}${day}_${hours}${minutes}${seconds}`;
-}
